refactor(websocket): doc NewDispatcher, defer unlock, race-gate stress test
Three follow-ups from Copilot review on PR #1065:

1. The doc comment promised concurrency safety but the zero value of
   Dispatcher has a nil handlers map and would panic on use. Make the
   constructor requirement explicit in the type doc.

2. Switch Register/RegisterFunc/HasHandler to defer-based unlock so a
   future early return inside the critical section does not silently
   leak the lock. Dispatch keeps manual unlock - releasing the read
   lock before invoking handler.Handle is intentional (serialising
   every dispatched message under the lock would risk deadlock with
   re-entrant Register and pin Handle behind unrelated I/O).

3. Split the concurrent stress test into handler_race_test.go with a
   //go:build race tag so it only runs under 'go test -race' (which
   the backend-tests CI workflow uses), avoiding the runtime overhead
   on plain 'go test' invocations.
