+++ apps/backend/pkg/websocket/handler.go
@@ -21,10 +21,11 @@ func (f HandlerFunc) Handle(ctx context.Context, msg *Message) (*Message, error)
 
 // Dispatcher routes messages to appropriate handlers based on action.
 //
-// Dispatcher is safe for concurrent use: Register/RegisterFunc may be
-// called from any goroutine, including while Dispatch or HasHandler are
-// in flight. Typical usage registers all handlers at startup, but the
-// API does not require it.
+// Dispatcher must be constructed via NewDispatcher; the zero value has a
+// nil handlers map and panics on use. Once constructed it is safe for
+// concurrent use: Register/RegisterFunc may be called from any goroutine,
+// including while Dispatch or HasHandler are in flight. Typical usage
+// registers all handlers at startup, but the API does not require it.
 type Dispatcher struct {
 	mu       sync.RWMutex
 	handlers map[string]Handler
@@ -40,18 +41,23 @@ func NewDispatcher() *Dispatcher {
 // Register registers a handler for an action
 func (d *Dispatcher) Register(action string, handler Handler) {
 	d.mu.Lock()
+	defer d.mu.Unlock()
 	d.handlers[action] = handler
-	d.mu.Unlock()
 }
 
 // RegisterFunc registers a handler function for an action
 func (d *Dispatcher) RegisterFunc(action string, handler HandlerFunc) {
 	d.mu.Lock()
+	defer d.mu.Unlock()
 	d.handlers[action] = handler
-	d.mu.Unlock()
 }
 
-// Dispatch routes a message to the appropriate handler
+// Dispatch routes a message to the appropriate handler.
+//
+// The handler is looked up under the read lock, then the lock is released
+// before invoking Handle: holding it across user-supplied I/O would
+// serialise every dispatched message and risk deadlock if a handler
+// re-entered Register.
 func (d *Dispatcher) Dispatch(ctx context.Context, msg *Message) (*Message, error) {
 	d.mu.RLock()
 	handler, ok := d.handlers[msg.Action]
@@ -66,7 +72,7 @@ func (d *Dispatcher) Dispatch(ctx context.Context, msg *Message) (*Message, erro
 // HasHandler returns true if a handler is registered for the action
 func (d *Dispatcher) HasHandler(action string) bool {
 	d.mu.RLock()
+	defer d.mu.RUnlock()
 	_, ok := d.handlers[action]
-	d.mu.RUnlock()
 	return ok
 }
