+++ apps/backend/pkg/websocket/handler_test.go
@@ -2,7 +2,6 @@ package websocket
 
 import (
 	"context"
-	"sync"
 	"testing"
 )
 
@@ -48,50 +47,4 @@ func TestDispatcher_UnknownActionReturnsError(t *testing.T) {
 		t.Errorf("expected error code %q, got %q",
 			ErrorCodeUnknownAction, payload.Code)
 	}
-}
-
-// TestDispatcher_ConcurrentRegisterAndDispatch exercises the Dispatcher
-// from many goroutines at once. The intent is that `go test -race` flags
-// any data race on the underlying handlers map; without the RWMutex this
-// test failed under -race with a "concurrent map read and map write"
-// report.
-func TestDispatcher_ConcurrentRegisterAndDispatch(t *testing.T) {
-	d := NewDispatcher()
-
-	const goroutines = 16
-	const ops = 200
-
-	noop := HandlerFunc(func(_ context.Context, msg *Message) (*Message, error) {
-		return &Message{ID: msg.ID, Action: msg.Action}, nil
-	})
-
-	var wg sync.WaitGroup
-	wg.Add(goroutines * 2)
-
-	for g := 0; g < goroutines; g++ {
-		go func() {
-			defer wg.Done()
-			for i := 0; i < ops; i++ {
-				// Alternate between Register and RegisterFunc so both
-				// write paths are exercised under -race.
-				if i%2 == 0 {
-					d.Register("action", noop)
-				} else {
-					d.RegisterFunc("action", noop)
-				}
-			}
-		}()
-		go func() {
-			defer wg.Done()
-			for i := 0; i < ops; i++ {
-				_, _ = d.Dispatch(context.Background(), &Message{
-					ID:     "x",
-					Action: "action",
-				})
-				_ = d.HasHandler("action")
-			}
-		}()
-	}
-
-	wg.Wait()
 }
